# springboot_openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9000
  description: Generated server url
paths:
...
  /managers/{id}/toggleAdmin:
    put:
      tags:
      - manager-controller
      operationId: toggleAdmin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /parks/{parkId}/vehicle:
    post:
      tags:
      - park-controller
      operationId: addVehicle
      parameters:
      - name: parkId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: vehicleType
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "404":
          description: Dock non trovata
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /parks/vehicles/move:
    post:
      tags:
      - park-controller
      operationId: getVehicles
      parameters:
      - name: vehicleIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      - name: parkId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          description: Not enough docks
          content:
            application/text:
              schema: {}
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /managers:
    get:
      tags:
      - manager-controller
      operationId: getAllManagers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOManagerOnlyWithParkCount"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
    post:
      tags:
      - manager-controller
      operationId: addManager
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DAOManagerInsert"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /parks:
    get:
      tags:
      - park-controller
      operationId: getParks
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOParkWithIdAndName"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /parks/{parkId}/vehicles:
    get:
      tags:
      - park-controller
      operationId: getVehiclesByPark
      parameters:
      - name: parkId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOVehicleWithStatus"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "404":
          description: Parcheggio non trovato
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /parks/{parkId}/report:
    get:
      tags:
      - park-controller
      operationId: getReportAI
      parameters:
      - name: parkId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DAOParkReportAI"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /parks/vehiclesWithStatusAndReports:
    get:
      tags:
      - park-controller
      summary: "Get parks with vehicles, status and reports"
      description: "Returns a list of parks with their docks and vehicles information including status and reports. For managers, returns only their assigned parks. For admins, returns all parks."
      operationId: getVehicles_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOParkWithVehicles"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "404":
          description: Username non trovato tra i Customer
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
  /parks/info:
    get:
      tags:
      - park-controller
      operationId: getParksInfo
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DAOParkWithVehicleTypeInfo"
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
...
  /managers/{id}:
    delete:
      tags:
      - manager-controller
      operationId: deleteManager
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: substitutorId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "403":
          description: Accesso negato per il ruolo fornito
          content:
            application/text:
              schema: {}
        "500":
          description: Errore interno del server
          content:
            application/text:
              schema: {}
components:
  schemas:
...
    DAOManagerInsert:
      required:
      - admin
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
        admin:
          type: boolean
...
    DAOParkWithIdAndName:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
    DAOVehicleWithStatus:
      required:
      - id
      - immissionDate
      - status
      - vehicleType
      type: object
      properties:
        id:
          type: integer
          format: int32
        immissionDate:
          type: string
          format: date
        dismissionDate:
          type: string
          format: date
        vehicleType:
          type: string
        status:
          type: string
          enum:
          - AVAILABLE
          - NOT_AVAILABLE
          - UNDER_MAINTENANCE
          - ABANDONED
          - DISABLED
    DAOParkReportAI:
      required:
      - priorities
      - state
      - summary
      type: object
      properties:
        state:
          type: integer
          format: int32
        summary:
          type: string
        priorities:
          type: array
          items:
            type: integer
            format: int32
    DAOCustomerReportOnly:
      required:
      - description
      - time
      type: object
      properties:
        time:
          type: string
          format: date-time
        description:
          type: string
        customer:
          type: string
    DAODockWithIdVehicle:
      required:
      - number
      type: object
      properties:
        number:
          type: integer
          format: int32
        idVehicle:
          type: integer
          format: int32
    DAOParkWithVehicles:
      required:
      - dockCount
      - docks
      - id
      - name
      - vehicles
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        dockCount:
          type: integer
          format: int32
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/DAOVehicleSmallWithStatus"
        docks:
          type: array
          items:
            $ref: "#/components/schemas/DAODockWithIdVehicle"
    DAOSensorReportOnly:
      required:
      - time
      - value
      type: object
      properties:
        time:
          type: string
          format: date-time
        value:
          type: number
          format: float
    DAOVehicleSmallWithStatus:
      required:
      - customerReports
      - id
      - sensorReports
      - status
      - vehicleType
      type: object
      properties:
        id:
          type: integer
          format: int32
        vehicleType:
          type: string
        status:
          type: boolean
        sensorReports:
          type: array
          items:
            $ref: "#/components/schemas/DAOSensorReportOnly"
        customerReports:
          type: array
          items:
            $ref: "#/components/schemas/DAOCustomerReportOnly"
    DAOParkWithVehicleTypeInfo:
      required:
      - bikeCount
      - ebikeCount
      - id
      - latitude
      - longitude
      - name
      - scooterCount
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        bikeCount:
          type: integer
          format: int32
        eBikeCount:
          type: integer
          format: int32
          writeOnly: true
        scooterCount:
          type: integer
          format: int32
        ebikeCount:
          type: integer
          format: int32
    DAOManagerOnlyWithParkCount:
      required:
      - admin
      - email
      - id
      - parkCount
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        admin:
          type: boolean
        parkCount:
          type: integer
          format: int32
...